+++ debug.h
@@ -9,7 +9,7 @@ struct cpu_driver;
 
 struct debug_struct* debug_create(struct bbc_struct* p_bbc,
                                   int debug_active,
-                                  uint16_t debug_stop_addr);
+                                  int32_t debug_stop_addr);
 void debug_destroy(struct debug_struct* p_debug);
 
 int debug_subsystem_active(void* p);
